backend/requirements.txt
django
djangorestframework
django-cors-headers
requests
python-dotenv
celery
django-celery-beat
django-filter
google-genai
